refactor: store.business 패키지 파일 이동 -> store.business.internal
