+++ common/src/main/java/com/food/common/store/business/internal/impl/DefaultStoreOwnerEntityService.java
@@ -1,6 +1,6 @@
-package com.food.common.store.business.impl;
+package com.food.common.store.business.internal.impl;
 
-import com.food.common.store.business.StoreOwnerEntityService;
+import com.food.common.store.business.internal.StoreOwnerEntityService;
 import com.food.common.store.domain.StoreOwner;
 import com.food.common.user.domain.User;
 import com.food.common.user.repository.StoreOwnerRepository;
